+++ CHANGELOG.md
@@ -11,6 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- if creation of incomplete time entries is not allowed, the commands will verify if the project is
+  active or not.
+- when closing a running time entry before creating a new one, the client will validate it before
+  asking information on interactive mode.
+
+### Fixed
+
+- archived projects were being shown as options to select in interactive mode, now only active are
+  shown.
+
 ## [v0.29.0] - 2022-01-12
 
 ### Changed